# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /recommendations:
    post:
      description: Loads profile answers and active packs/products/references, selects
        best references, stores the recommendation session/results/items, and
        returns ranked packs with score details.
      operationId: RecommendationsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationDto"
      responses:
        "201":
          description: Recommendations generated and stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "400":
          description: Profile has no answers or request validation failed.
        "404":
          description: Customer profile not found.
      summary: Generate recommendations for a customer profile
      tags: &a34
        - Recommendations
  /recommendations/{sessionId}:
    get:
      description: Returns a previously generated recommendation session and ranked results.
      operationId: RecommendationsController_findOne
      parameters:
        - name: sessionId
          required: true
          in: path
          description: Recommendation session ID.
          schema:
            example: 00000000-0000-4000-8000-000000000001
            type: string
      responses:
        "200":
          description: Stored recommendation session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "404":
          description: Recommendation session not found.
      summary: Get a stored recommendation session
      tags: *a34
  /admin/recommendation-rules:
    get:
      description: OWNER, ADMIN, and STAFF can read paginated recommendation scoring rules.
      operationId: AdminRecommendationRulesController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: skin
            type: string
        - name: targetType
          required: false
          in: query
          schema:
            type: string
            enum:
              - PACK
              - PRODUCT
              - REFERENCE
        - name: conditionType
          required: false
          in: query
          schema:
            type: string
            enum:
              - MUST_MATCH
              - SHOULD_MATCH
              - EXCLUDE_IF_MATCH
        - name: attributeGroupId
          required: false
          in: query
          schema:
            example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
            type: string
        - name: isActive
          required: false
          in: query
          schema:
            example: true
            type: boolean
      responses:
        "200":
          description: Paginated recommendation rules.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationRuleResponse"
        "401": &a35
          description: Authentication required.
        "403": &a36
          description: Insufficient admin role.
      security: &a37
        - bearer: []
      summary: List recommendation rules
      tags: &a38
        - Admin Recommendation Rules
    post:
      description: OWNER and ADMIN can create scoring rules. Rule codes are immutable
        after creation.
      operationId: AdminRecommendationRulesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationRuleDto"
      responses:
        "200":
          description: Recommendation rule created.
        "401": *a35
        "403": *a36
        "409":
          description: Recommendation rule code already exists.
      security: *a37
      summary: Create recommendation rule
      tags: *a38
  /admin/recommendation-rules/preview:
    post:
      description: OWNER, ADMIN, and STAFF can preview current rule effects. This
        endpoint does not create recommendation sessions, results, or items.
      operationId: AdminRecommendationRulesController_preview
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationDto"
      responses:
        "200":
          description: Preview recommendations without persisting a session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "401": *a35
        "403": *a36
      security: *a37
      summary: Preview recommendations without persistence
      tags: *a38
  /admin/recommendation-rules/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can read recommendation rule details.
      operationId: AdminRecommendationRulesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Recommendation rule details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationRuleResponse"
        "401": *a35
        "403": *a36
        "404":
          description: Recommendation rule not found.
      security: *a37
      summary: Get recommendation rule details
      tags: *a38
    patch:
      description: OWNER and ADMIN can update mutable recommendation rule fields.
      operationId: AdminRecommendationRulesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecommendationRuleDto"
      responses:
        "200":
          description: Recommendation rule updated.
        "401": *a35
        "403": *a36
        "404":
          description: Recommendation rule not found.
      security: *a37
      summary: Update recommendation rule
      tags: *a38
    delete:
      description: OWNER and ADMIN can soft-deactivate recommendation rules.
      operationId: AdminRecommendationRulesController_deactivate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Recommendation rule deactivated.
        "401": *a35
        "403": *a36
        "404":
          description: Recommendation rule not found.
      security: *a37
      summary: Deactivate recommendation rule
      tags: *a38
...
info:
  title: Beauty Pack Recommendation API
  description: Backend API for beauty-pack personalization, recommendations, COD
    orders, and administration
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the JWT access token returned by POST /auth/login.
  schemas:
...
    AttributeOptionResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: MEDIUM
        label:
          type: string
          example: Medium
      required:
        - id
        - code
        - label
...
    QuizAttributeGroupResponse:
      type: object
      properties:
        code:
          type: string
          example: SKIN_COLOR
        name:
          type: string
          example: Skin Color
      required:
        - code
        - name
...
    CategorySummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: FOUNDATION
        name:
          type: string
          example: Foundation
      required:
        - id
        - code
        - name
    BrandSummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Demo Beauty
      required:
        - id
        - name
    AttributeMatchResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        matchType:
          type: string
          enum:
            - COMPATIBLE
            - NOT_COMPATIBLE
            - BOOST
          example: COMPATIBLE
        scoreValue:
          type: number
          example: 40
        isHardFilter:
          type: boolean
          example: false
        attributeGroup:
          $ref: "#/components/schemas/QuizAttributeGroupResponse"
        attributeOption:
          $ref: "#/components/schemas/AttributeOptionResponse"
      required:
        - id
        - matchType
        - scoreValue
        - isHardFilter
        - attributeGroup
        - attributeOption
    ProductReferenceResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        referenceCode:
          type: string
          example: RF2
        referenceName:
          type: string
          example: Medium Warm
        sku:
          type: object
          example: FOUNDATION-X-RF2
          nullable: true
        barcode:
          type: object
          example: null
          nullable: true
        priceDelta:
          type: number
          example: 0
        stockQuantity:
          type: number
          example: 20
        reservedQuantity:
          type: number
          example: 0
        isActive:
          type: boolean
          example: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/AttributeMatchResponse"
      required:
        - id
        - referenceCode
        - referenceName
        - priceDelta
        - stockQuantity
        - reservedQuantity
        - isActive
        - attributes
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Foundation X
        slug:
          type: string
          example: foundation-x
        basePrice:
          type: number
          example: 120
        currency:
          type: string
          example: MAD
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - ARCHIVED
          example: ACTIVE
        category:
          $ref: "#/components/schemas/CategorySummaryResponse"
        brand:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/BrandSummaryResponse"
        references:
          type: array
          items:
            $ref: "#/components/schemas/ProductReferenceResponse"
      required:
        - id
        - name
        - slug
        - basePrice
        - currency
        - status
        - category
        - references
...
    PackItemResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        selectionMode:
          type: string
          enum:
            - FIXED_REFERENCE
            - AUTO_BEST_REFERENCE
            - CUSTOMER_CHOICE
          example: AUTO_BEST_REFERENCE
        quantity:
          type: number
          example: 1
        isRequired:
          type: boolean
          example: true
        product:
          $ref: "#/components/schemas/ProductResponse"
        productReference:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/ProductReferenceResponse"
      required:
        - id
        - selectionMode
        - quantity
        - isRequired
        - product
    PackResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Natural Glow Pack
        slug:
          type: string
          example: natural-glow-pack
        priceMode:
          type: string
          enum:
            - FIXED
            - SUM_ITEMS
            - SUM_ITEMS_WITH_DISCOUNT
          example: FIXED
        fixedPrice:
          type: object
          example: 299
          nullable: true
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - ARCHIVED
          example: ACTIVE
        isActive:
          type: boolean
          example: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/AttributeMatchResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PackItemResponse"
      required:
        - id
        - name
        - slug
        - priceMode
        - status
        - isActive
        - attributes
        - items
...
    CreateRecommendationDto:
      type: object
      properties:
        customerProfileId:
          type: string
          description: Customer profile ID returned by POST /quiz/profiles.
          example: 00000000-0000-4000-8000-000000000001
      required:
        - customerProfileId
    RecommendationResultItemResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        quantity:
          type: number
          example: 1
        itemScore:
          type: number
          example: 65
        product:
          $ref: "#/components/schemas/ProductResponse"
        selectedProductReference:
          $ref: "#/components/schemas/ProductReferenceResponse"
      required:
        - id
        - quantity
        - itemScore
        - product
        - selectedProductReference
    RecommendationPackResponse:
      type: object
      properties:
        recommendationResultId:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        rank:
          type: number
          example: 1
        totalScore:
          type: number
          example: 83
        matchPercentage:
          type: number
          example: 83
        reasonSummary:
          type: string
          example: Strong match for your selected profile.
        pack:
          $ref: "#/components/schemas/PackResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationResultItemResponse"
        reasonJson:
          type: object
          example:
            packScore: 30
            rawItemsScore: 145
            normalizedItemsScore: 48
            priorityBonus: 5
      required:
        - recommendationResultId
        - rank
        - totalScore
        - matchPercentage
        - reasonSummary
        - pack
        - items
        - reasonJson
    RecommendationResponse:
      type: object
      properties:
        sessionId:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        status:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - COMPLETED
            - FAILED
          example: COMPLETED
        algorithmVersion:
          type: string
          example: v1
        recommendedPacks:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationPackResponse"
      required:
        - sessionId
        - status
        - algorithmVersion
        - recommendedPacks
    RecommendationRuleResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: SKIN_COLOR_MATCH
        name:
          type: string
          example: Skin color match
        targetType:
          type: string
          enum:
            - PACK
            - PRODUCT
            - REFERENCE
        conditionType:
          type: string
          enum:
            - MUST_MATCH
            - SHOULD_MATCH
            - EXCLUDE_IF_MATCH
        scoreValue:
          type: number
          example: 40
        weight:
          type: number
          example: 1
        isActive:
          type: boolean
          example: true
      required:
        - id
        - code
        - name
        - targetType
        - conditionType
        - scoreValue
        - weight
        - isActive
    CreateRecommendationRuleDto:
      type: object
      properties:
        code:
          type: string
          example: COVERAGE_MATCH
        name:
          type: string
          example: Coverage match
        targetType:
          type: string
          enum: &a49
            - PACK
            - PRODUCT
            - REFERENCE
        attributeGroupId:
          type: object
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        conditionType:
          type: string
          enum: &a50
            - MUST_MATCH
            - SHOULD_MATCH
            - EXCLUDE_IF_MATCH
        scoreValue:
          type: number
          example: 15
        weight:
          type: number
          example: 1
          default: 1
        isActive:
          type: boolean
          example: true
          default: true
      required:
        - code
        - name
        - targetType
        - conditionType
        - scoreValue
    UpdateRecommendationRuleDto:
      type: object
      properties:
        name:
          type: string
          example: Coverage match
        targetType:
          type: string
          enum: *a49
        attributeGroupId:
          type: object
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        conditionType:
          type: string
          enum: *a50
        scoreValue:
          type: number
          example: 15
        weight:
          type: number
          example: 1
          default: 1
        isActive:
          type: boolean
          example: true
          default: true
...